apps/catalog/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
SQLAlchemy
